rename `parse` to `attrs`
